apps/backend/agent_in_update/requirements.txt
mangum
fastapi
uvicorn
langgraph
langchain-groq
openai
amadeus
httpx
python-multipart
requests
python-dotenv
nest-asyncio
pydantic